updated: java docs
warning: unsafe to use

v0.0.4
