+++ src/Main.java
@@ -1,5 +1,5 @@
 public class Main {
     public static void main(String[] args) {
-        utils.window.MainWindow.startVAGAI();
+        utils.mainWindow.MainWindow.startVAGAI();
     }
 }
